rename execution_loop -> execution_handlers
